+++ german_accounting/hooks.py
@@ -69,9 +69,6 @@ doc_events = {
     "Sales Invoice": {
         "validate": "german_accounting.events.extended_tax_category.validate_tax_category_fields",
 	},
-	"Customer": {
-		"onload": "german_accounting.events.update_amounts_value.update_amounts"
-	}
 }
 # doctype_list_js = {"doctype" : "public/js/doctype_list.js"}
 # doctype_tree_js = {"doctype" : "public/js/doctype_tree.js"}
