+++ german_accounting/public/js/banner/quotation_banner.js
@@ -1,12 +1,8 @@
+frappe.require('/assets/german_accounting/js/banner_utils.js')
 frappe.ui.form.on('Quotation', {
     party_name: async (frm) => {
         if(frm.doc.party_name) {
-            await updateAmounts(frm)
-        }
-    },
-    onload: async (frm) => {
-        if(frm.doc.party_name) {
-            await updateAmounts(frm)
+            frm.trigger('make_dashboard');
         }
     },
 	refresh: (frm) => {
@@ -15,226 +11,52 @@ frappe.ui.form.on('Quotation', {
 	make_dashboard: async (frm) => {
 		
             if (frm.doc.party_name) {
+                const customer = frm.doc.party_name
+                const company = frm.doc.company
                 const doctype = frm.doc.doctype
-                const currencySymbol = await getDefaultCurrencySymbol();
-                const credit_limit =  (await getCreditLimit(frm.doc.party_name, frm.doc.company, doctype)).toFixed(2);
-                const bypass_checked = await check_bypass(frm, doctype);
+                const docname = frm.doc.name
 
-                let customer = frm.doc.party_name
-                let open_invoice_amount = frm.doc.open_invoice_amount.toFixed(2)
-                let overdue_invoice_amount = frm.doc.overdue_invoice_amount.toFixed(2)
-                let non_invoiced_amount = frm.doc.non_invoiced_amount.toFixed(2)
-                let total = (parseFloat(open_invoice_amount) + parseFloat(non_invoiced_amount)).toFixed(2)
+                const currencySymbol = await getCurrencySymbol()
+                const credit_limit =  (await getCreditLimit(customer, company, doctype)).toFixed(2);
+                const bypass_checked = await checkBypass(customer, company, doctype);
+                const amounts = await updateAmounts(customer)
+                
+
+                const open_invoice_amount = amounts.open_invoice.toFixed(2)
+                const overdue_invoice_amount = amounts.overdue_invoice.toFixed(2)
+                const non_invoiced_amount = amounts.non_invoice.toFixed(2)
+                const total = amounts.total.toFixed(2)
 
                 let creditLimitText = credit_limit !== '0.00' ?
                 `${__('and The Credit Limit is')}: <b>${currencySymbol} ${credit_limit}</b>` :
                 `<br> <span style="color: #ff4d4d;">${__('Credit Limit is not set for this customer')}</span>`;
 
-                let textColor = '#1366AE'; // Default text color
+                let textClass = '';
                 if ((parseFloat(total) > parseFloat(credit_limit)) && parseFloat(credit_limit) > 0) {
                     if (!bypass_checked){
                         $('button[data-label="Submit"]').off()
-                        $('button[data-label="Submit"]').click(() => {check_credit_limit(frm, doctype)});
+                        $('button[data-label="Submit"]').click(() => {checkCreditLimit(frm, customer, company, doctype, docname, total)});
                     }
-                    textColor = '#ff4d4d'; // Red text color
+                    textClass = 'text-danger';
                 }
                 frm.dashboard.clear_headline();
 
                 frm.dashboard.set_headline_alert(`
                     <div class="row">
                         <div class="col-xs-12">
-                            <span class="indicator whitespace-nowrap" style="color: ${textColor};">
+                            <span class="indicator whitespace-nowrap ${textClass}">
                                 <span class="hidden-xs">
-                                    <p>${__('Customer')} <b>${customer}</b> ${__('has an Open Invoice Amount of')}: <b>${currencySymbol} ${open_invoice_amount}</b>, ${__('Overdue Invoice Amount of')}: <b>${currencySymbol} ${overdue_invoice_amount}</b> ${__('and Non-Invoiced Amount of')}: <b>${currencySymbol} ${non_invoiced_amount}</b><br> ${__('This is a Total of')}: <b>${currencySymbol} ${total}</b> ${creditLimitText}
+                                    <p>
+                                        ${__("Customer {} has an Open Invoice Amount of: {}, Overdue Invoice Amount of: {} and Non-Invoiced Amount of: {}",[`<b>${customer}</b>`, `<b>${currencySymbol} ${open_invoice_amount}</b>`, `<b>${currencySymbol} ${overdue_invoice_amount}</b>`, `<b>${currencySymbol} ${non_invoiced_amount}</b>`])}
+                                        <br>
+                                        ${__("This is a Total of {}", [`<b>${currencySymbol} ${total}</b>`])}
+                                        ${creditLimitText}
                                     </p>
                                 </span>
                             </span>
                         </div>
                     </div>
-                `);
-                
-            
+                `);            
             }
 	}
 })
-
-
-async function updateAmounts(frm) {
-
-    const customer = frm.doc.party_name
-    frappe.call({
-        method: "german_accounting.events.sales_order_amount.update_amounts",
-        args: {
-            customer: customer
-        },
-        callback: function (r) {
-           const val = r.message
-           frm.doc.open_invoice_amount = val[0]
-           frm.doc.overdue_invoice_amount = val[1]
-           frm.doc.non_invoiced_amount = val[2]
-           frm.doc.total_outstanding_amount = val[3]
-        },
-    });
-}
-
-
-function check_credit_limit(frm, doctype) {
-    const docname = frm.doc.name;
-    const customer = frm.doc.party_name;
-    const company = frm.doc.company;
-    const total = frm.doc.total_outstanding_amount;
-
-    frappe.call({
-        method: "german_accounting.events.credit_limit_check.check_credit_limit",
-        args: {
-            docname,
-            customer,
-            company,
-            total,
-            doctype
-        },
-        callback: function (r) {
-            const response = r.message;
-           
-            const dialog = new frappe.ui.Dialog({
-                title: ('Are you sure you want to proceed?'),
-                fields: [
-                    {
-                        fieldtype: 'HTML',
-                        fieldname: 'message',
-                        options: `<p>${response.message}</p>`
-                    },
-                    {
-                        fieldtype: 'HTML',
-                        fieldname: 'users',
-                        options: response.users
-                    }
-                ],
-                primary_action_label: (response.button_label),
-                primary_action: function() {
-                    if (response.button_label == "Submit"){
-
-                        frm.save("Submit");
-
-                    } else {
-
-                        const selectedUsers = [];
-                        $(dialog.$wrapper).find('input[name="user_checkbox"]:checked').each(function() {
-                            selectedUsers.push($(this).val());
-                        });
-                        
-                        // send email
-                        frappe.call({
-                            method: "german_accounting.events.credit_limit_check.send_emails",
-                            args: {
-                                users: selectedUsers,
-                                docname: docname,
-                                doctype
-                            },
-                            callback: function (r) {
-
-                                if (r.message) {
-                                    if (r.message.status === "success") {
-                                        frappe.msgprint({
-                                            title: __('Success'),
-                                            indicator: 'green',
-                                            message: r.message.message
-                                        });
-                                    } else {
-                                        frappe.msgprint({
-                                            title: __('Error'),
-                                            indicator: 'red',
-                                            message: r.message.message
-                                        });
-                                    }
-                                }
-                            }
-                        })
-                    }
-
-                    dialog.hide();
-                },
-                secondary_action_label: __('Cancel'),
-                secondary_action: function() {
-                    dialog.hide();
-                }
-            });
-
-            dialog.show();
-
-            $(dialog.$wrapper).find('#select-all').on('change', function() {
-                const isChecked = $(this).is(':checked');
-                $(dialog.$wrapper).find('input[name="user_checkbox"]').prop('checked', isChecked);
-            });
-
-            $(dialog.$wrapper).find('.modal-footer .btn-primary').addClass('btn-danger').removeClass('btn-primary');
-        }
-    });
-} 
-
-async function getDefaultCurrencySymbol() {
-    try {
-
-        const response = await fetch("/api/resource/Global Defaults/Global Defaults");
-        const data = await response.json();
-
-        const currencyCode = data.data.default_currency;
-
-        const currencyResponse = await fetch(`/api/resource/Currency/${currencyCode}`);
-        const currencyData = await currencyResponse.json();
-
-        const currencySymbol = currencyData.data.symbol;
-
-        return currencySymbol;
-    } catch (error) {
-        console.error("Error fetching default currency symbol:", error);
-        return null;
-    }
-}
-
-
-async function getCreditLimit(customer, company, doctype) {
-    return new Promise((resolve, reject) => {
-        frappe.call({
-            method: 'german_accounting.events.credit_limit_check.get_credit_limit',
-            args: {
-                customer,
-                company,
-                doctype
-            },
-            callback: function(r) {
-                if (r.message !== null && r.message !== undefined) {
-                    resolve(r.message);
-                } else {
-                    resolve(0);
-                }
-            }
-        });
-    });
-}
-
-
-
-async function check_bypass(frm, doctype) {
-    return new Promise((resolve, reject) => {
-
-        const customer = frm.doc.party_name;
-        const company = frm.doc.company;
-
-        frappe.call({
-            method: 'german_accounting.events.credit_limit_check.bypass_checked',
-            args: {
-                customer,
-                company,
-                doctype
-            },
-            callback: function(r) {
-                if (r.message !== null && r.message !== undefined) {
-                    resolve(r.message);
-                } else {
-                    reject(new Error('Response not found'));
-                }
-            }
-        });
-    })
-}
