+++ german_accounting/translations/de.csv
@@ -1,3 +1,3 @@
 Sales Invoice Type,Rechnungstyp
 Invoice Cancellation,Stornorechnung
-Credit Note,Bonusgutschrift
+Credit Note,Gutschrift
